Registering MetadataStore in constructor instead of StoreConfig
