+++ src/js/stores/MetadataStore.js
@@ -1,44 +1,64 @@
 import deepEqual from 'deep-equal';
-import GetSetBaseStore from './GetSetBaseStore';
 
-var AppDispatcher = require('../events/AppDispatcher');
-import ActionTypes from '../constants/ActionTypes';
+import {
+  REQUEST_DCOS_METADATA,
+  REQUEST_METADATA,
+  SERVER_ACTION
+} from '../constants/ActionTypes';
+import AppDispatcher from '../events/AppDispatcher';
 import Config from '../config/Config';
-import EventTypes from '../constants/EventTypes';
+import {
+  DCOS_METADATA_CHANGE,
+  METADATA_CHANGE
+} from '../constants/EventTypes';
+import GetSetBaseStore from './GetSetBaseStore';
 import MetadataActions from '../events/MetadataActions';
+import PluginSDK from 'PluginSDK';
 
 class MetadataStore extends GetSetBaseStore {
-
   constructor() {
     super(...arguments);
 
+    PluginSDK.addStoreConfig({
+      store: this,
+      storeID: this.storeID,
+      events: {
+        success: METADATA_CHANGE,
+        dcosSuccess: DCOS_METADATA_CHANGE
+      },
+      unmountWhen: function () {
+        return true;
+      },
+      listenAlways: true
+    });
+
     this.dispatcherIndex = AppDispatcher.register((payload) => {
       var source = payload.source;
-      if (source !== ActionTypes.SERVER_ACTION) {
+      if (source !== SERVER_ACTION) {
         return false;
       }
 
       var action = payload.action;
 
       switch (action.type) {
-        case ActionTypes.REQUEST_METADATA:
+        case REQUEST_METADATA:
           var oldMetadata = this.get('metadata');
           var metadata = action.data;
 
           // only emitting on change
           if (!deepEqual(oldMetadata, metadata)) {
             this.set({metadata});
-            this.emitChange(EventTypes.METADATA_CHANGE);
+            this.emitChange(METADATA_CHANGE);
           }
           break;
-        case ActionTypes.REQUEST_DCOS_METADATA:
+        case REQUEST_DCOS_METADATA:
           var oldDCOSMetadata = this.get('dcosMetadata');
           var dcosMetadata = action.data;
 
           // only emitting on change
           if (!deepEqual(oldDCOSMetadata, dcosMetadata)) {
             this.set({dcosMetadata});
-            this.emitChange(EventTypes.DCOS_METADATA_CHANGE);
+            this.emitChange(DCOS_METADATA_CHANGE);
           }
           break;
       }
